refactor(a11y-tests): move story testing helpers to dedicated file
