+++ eslint.config.js
@@ -62,7 +62,8 @@ export default [
       'rollup/**/*.js',
       'cem/**/*.js',
       'web-test-runner.config.js',
-      'wds/**/*.js',
+      'web-test-runner.config*.js',
+      'web-test-runner/**/*.js',
       'src/stories/lib/smart-auth-plugin.js',
       'test-mocha/**/*.*js',
       '.storybook/**/*.js',
@@ -90,7 +91,8 @@ export default [
       'rollup/**/*.js',
       'cem/**/*.js',
       'web-test-runner.config.js',
-      'wds/**/*.js',
+      'web-test-runner.visual-tests.config.js',
+      'web-test-runner/**/*.js',
       '.storybook/**/*.js',
     ],
     rules: {
