+++ test/helpers/story-testing-utils.js
@@ -1,0 +1,68 @@
+import { ignoreWindowOnError } from '@lit-labs/virtualizer/support/resize-observer-errors.js';
+
+/**
+ * @typedef {import('./story-testing-utils.types.js').RawStoriesModule} RawStoriesModule
+ * @typedef {import('./story-testing-utils.types.js').AnnotatedStoryFnClever} AnnotatedStoryFnClever
+ * @typedef {import('./story-testing-utils.types.js').ModuleEntryWithStoryFunction} ModuleEntryWithStoryFunction
+ * @typedef {import('./story-testing-utils.types.js').ArrayOfStoryFunctions} ArrayOfStoryFunctions
+ * @typedef {import('@storybook/web-components').WebComponentsRenderer} WebComponentsRenderer
+ * @typedef {import('@storybook/web-components').StoryContext<WebComponentsRenderer>} StoryContext
+ */
+
+export const viewports = {
+  desktop: {
+    width: 1200,
+    height: 640,
+  },
+  mobile: {
+    width: 375,
+    height: 640,
+  },
+};
+
+/** @type {StoryContext} */
+// @ts-ignore we don't need to provide all options because the story won't be displayed in storybook
+export const storyConf = {
+  globals: {
+    locale: 'en',
+  },
+};
+
+/**
+ * Sets up error handling to ignore specific errors during test execution
+ *
+ * @param {Function} before - Function to execute before test
+ * @param {Function} after - Function to execute after test
+ * @param {(message: string) => boolean} messagePredicate - Function that determines which error messages to ignore
+ */
+export function setupIgnoreIrrelevantErrors(before, after, messagePredicate) {
+  /** @type {Function|null} */
+  let teardown;
+  before(() => {
+    teardown = ignoreWindowOnError(messagePredicate);
+  });
+  after(() => {
+    teardown?.();
+    teardown = undefined;
+  });
+}
+
+/**
+ * Transform the result of an imported module from a story file into an array of story functions that can be used to render every story.
+ * The imported module is an object containing every named export (story functions mostly) as well as the default export (metadata about the stories).
+ *
+ * @param {RawStoriesModule} importedModule the result of a story file import.
+ * @returns {ArrayOfStoryFunctions} the story functions. These functions can be used to render each story.
+ */
+export const getStories = (importedModule) => {
+  const filteredStories = /** @type {ArrayOfStoryFunctions} */ (
+    Object.entries(importedModule)
+      .filter(
+        ([_moduleEntryName, moduleEntryValue]) =>
+          typeof moduleEntryValue === 'function' && 'component' in moduleEntryValue,
+      )
+      .map(([storyName, storyFunction]) => ({ storyName, storyFunction }))
+  );
+
+  return filteredStories;
+};
