+++ web-test-runner/story-file-to-a11y-tests-file-plugin.js
@@ -1,5 +1,6 @@
-export const testStoriesPlugin = {
-  name: 'test-story',
+/** @type {import('@web/test-runner').TestRunnerPlugin} */
+export const storyFileToA11yTestsFilePlugin = {
+  name: 'story-file-to-a11y-tests-file',
   async transformImport({ source, context }) {
     // if `.stories.js` is imported by WTR itself, then we change it to import the test file
     if (context.request.url.startsWith('/?wtr-session-id=') && source.includes('.stories.js?wtr-session')) {
@@ -10,10 +11,10 @@ export const testStoriesPlugin = {
     // test files are generated on the fly and they import story modules
     if (context.path.endsWith('.stories.test.js')) {
       const testFileContent = `
-        import { testStories } from '/test/helpers/test-stories.js';
+        import { runA11yTests } from '/test/helpers/a11y-tests.js';
         import * as storiesModule from '${context.path.replace('.stories.test.js', '.stories.js')}';
 
-        testStories(storiesModule);
+        runA11yTests(storiesModule);
       `;
       return testFileContent;
     }
